Add `configFile` option
This is passed through to `pg_format`.

It should be noted that `pg_format` looks in `~/.pg_format` if it exists
unless you tell it not to do so.  I couldn't find anywhere this was done
in this lib, so adding that file *should* work for i.e. the vscode
extension.  I haven't tested this, but it would be a great alternative
to allow a single place for configuraation.